requirements.txt
flask
openai
requests
python-dotenv
apscheduler
